doc-qwen3.5-122b-a10b/backend/requirements.txt
# ── Core backend ──────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── PDF processing ─────────────────────────────────────────
pymupdf
pillow

# ── Model client HTTP (used only on GPU machine) ──────────
# httpx already installed as fastapi dependency — no extra install needed

# ── GPU machine only — install these on workstation ────────
# torch --index-url https://download.pytorch.org/whl/cu124
# transformers[serving] @ git+https://github.com/huggingface/transformers.git@main
# torchvision
# accelerate
# huggingface_hub
